requirements.txt
streamlit
openai
llama-index==0.8.57
nltk
pypdf
langchain
pinecone-client
